refactor org.apache.shardingsphere.driver.spring.namespace.parser.rule
