+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-core-spring/shardingsphere-jdbc-core-spring-namespace/src/main/java/org/apache/shardingsphere/driver/spring/namespace/handler/SPINamespaceHandler.java
@@ -17,8 +17,8 @@
 
 package org.apache.shardingsphere.driver.spring.namespace.handler;
 
-import org.apache.shardingsphere.driver.spring.namespace.parser.rule.KeyGenerateAlgorithmBeanDefinitionParser;
-import org.apache.shardingsphere.driver.spring.namespace.parser.rule.ShardingAlgorithmBeanDefinitionParser;
+import org.apache.shardingsphere.driver.spring.namespace.parser.rule.sharding.KeyGenerateAlgorithmBeanDefinitionParser;
+import org.apache.shardingsphere.driver.spring.namespace.parser.rule.sharding.ShardingAlgorithmBeanDefinitionParser;
 import org.apache.shardingsphere.driver.spring.namespace.constants.SPIBeanDefinitionParserTag;
 import org.springframework.beans.factory.xml.NamespaceHandlerSupport;
