+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-core-spring/shardingsphere-jdbc-core-spring-namespace/src/main/java/org/apache/shardingsphere/driver/spring/namespace/parser/rule/masterslave/MasterSlaveDataSourceConfigurationBeanDefinition.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.driver.spring.namespace.parser.rule;
+package org.apache.shardingsphere.driver.spring.namespace.parser.rule.masterslave;
 
 import com.google.common.base.Splitter;
 import com.google.common.base.Strings;
