+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-core-spring/shardingsphere-jdbc-core-spring-namespace/src/main/java/org/apache/shardingsphere/driver/spring/namespace/parser/rule/masterslave/MasterSlaveLoadBalanceStrategyBeanDefinitionParser.java
@@ -15,33 +15,34 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.driver.spring.namespace.parser.rule;
+package org.apache.shardingsphere.driver.spring.namespace.parser.rule.masterslave;
 
 import com.google.common.base.Strings;
-import java.util.Properties;
-import org.apache.shardingsphere.driver.spring.namespace.constants.SPIBeanDefinitionParserTag;
-import org.apache.shardingsphere.driver.spring.namespace.factorybean.KeyGenerateAlgorithmFactoryBean;
+import org.apache.shardingsphere.masterslave.api.config.LoadBalanceStrategyConfiguration;
+import org.apache.shardingsphere.driver.spring.namespace.constants.LoadBalanceAlgorithmBeanDefinitionParserTag;
 import org.springframework.beans.factory.support.AbstractBeanDefinition;
 import org.springframework.beans.factory.support.BeanDefinitionBuilder;
 import org.springframework.beans.factory.xml.AbstractBeanDefinitionParser;
 import org.springframework.beans.factory.xml.ParserContext;
 import org.w3c.dom.Element;
 
+import java.util.Properties;
+
 /**
- * Key generate algorithm bean parser for spring namespace.
+ * Master slave load balance strategy bean parser for spring namespace.
  */
-public final class KeyGenerateAlgorithmBeanDefinitionParser extends AbstractBeanDefinitionParser {
+public final class MasterSlaveLoadBalanceStrategyBeanDefinitionParser extends AbstractBeanDefinitionParser {
     
     @Override
     protected AbstractBeanDefinition parseInternal(final Element element, final ParserContext parserContext) {
-        BeanDefinitionBuilder factory = BeanDefinitionBuilder.rootBeanDefinition(KeyGenerateAlgorithmFactoryBean.class);
-        factory.addConstructorArgValue(element.getAttribute(SPIBeanDefinitionParserTag.KEY_GENERATE_ALGORITHM_TYPE_ATTRIBUTE));
+        BeanDefinitionBuilder factory = BeanDefinitionBuilder.rootBeanDefinition(LoadBalanceStrategyConfiguration.class);
+        factory.addConstructorArgValue(element.getAttribute(LoadBalanceAlgorithmBeanDefinitionParserTag.ALGORITHM_TYPE_ATTRIBUTE));
         parseProperties(element, factory);
         return factory.getBeanDefinition();
     }
     
     private void parseProperties(final Element element, final BeanDefinitionBuilder factory) {
-        String properties = element.getAttribute(SPIBeanDefinitionParserTag.KEY_GENERATE_ALGORITHM_PROPERTY_REF_ATTRIBUTE);
+        String properties = element.getAttribute(LoadBalanceAlgorithmBeanDefinitionParserTag.ALGORITHM_PROPERTY_REF_ATTRIBUTE);
         if (!Strings.isNullOrEmpty(properties)) {
             factory.addConstructorArgReference(properties);
         } else {
