+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-core-spring/shardingsphere-jdbc-core-spring-namespace/src/main/java/org/apache/shardingsphere/driver/spring/namespace/parser/rule/sharding/KeyGeneratorBeanDefinitionParser.java
@@ -15,20 +15,26 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.driver.spring.namespace.parser.rule;
+package org.apache.shardingsphere.driver.spring.namespace.parser.rule.sharding;
 
+import org.apache.shardingsphere.sharding.api.config.KeyGeneratorConfiguration;
+import org.apache.shardingsphere.driver.spring.namespace.constants.ShardingRuleBeanDefinitionParserTag;
 import org.springframework.beans.factory.support.AbstractBeanDefinition;
+import org.springframework.beans.factory.support.BeanDefinitionBuilder;
 import org.springframework.beans.factory.xml.AbstractBeanDefinitionParser;
 import org.springframework.beans.factory.xml.ParserContext;
 import org.w3c.dom.Element;
 
 /**
- * Sharding strategy bean parser for spring namespace.
+ * Key generator bean parser for spring namespace.
  */
-public final class ShardingStrategyBeanDefinitionParser extends AbstractBeanDefinitionParser {
+public final class KeyGeneratorBeanDefinitionParser extends AbstractBeanDefinitionParser {
     
     @Override
     protected AbstractBeanDefinition parseInternal(final Element element, final ParserContext parserContext) {
-        return ShardingStrategyBeanDefinition.getBeanDefinitionByElement(element);
+        BeanDefinitionBuilder factory = BeanDefinitionBuilder.rootBeanDefinition(KeyGeneratorConfiguration.class);
+        factory.addConstructorArgValue(element.getAttribute(ShardingRuleBeanDefinitionParserTag.GENERATE_KEY_COLUMN_ATTRIBUTE));
+        factory.addConstructorArgReference(element.getAttribute(ShardingRuleBeanDefinitionParserTag.GENERATE_KEY_ALGORITHM_REF_TAG));
+        return factory.getBeanDefinition();
     }
 }
