+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-core-spring/shardingsphere-jdbc-core-spring-namespace/src/main/java/org/apache/shardingsphere/driver/spring/namespace/parser/rule/sharding/ShardingAlgorithmBeanDefinitionParser.java
@@ -15,28 +15,38 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.driver.spring.namespace.parser.rule;
+package org.apache.shardingsphere.driver.spring.namespace.parser.rule.sharding;
 
-import org.apache.shardingsphere.driver.spring.transaction.ShardingTransactionTypeScanner;
+import com.google.common.base.Strings;
+import org.apache.shardingsphere.driver.spring.namespace.constants.SPIBeanDefinitionParserTag;
+import org.apache.shardingsphere.driver.spring.namespace.factorybean.ShardingAlgorithmFactoryBean;
 import org.springframework.beans.factory.support.AbstractBeanDefinition;
 import org.springframework.beans.factory.support.BeanDefinitionBuilder;
 import org.springframework.beans.factory.xml.AbstractBeanDefinitionParser;
 import org.springframework.beans.factory.xml.ParserContext;
 import org.w3c.dom.Element;
 
+import java.util.Properties;
+
 /**
- * Sharding transaction type scanner bean definition parser.
+ * Sharding algorithm bean parser for spring namespace.
  */
-public final class ShardingTransactionTypeScannerBeanDefinitionParser extends AbstractBeanDefinitionParser {
-    
+public final class ShardingAlgorithmBeanDefinitionParser extends AbstractBeanDefinitionParser {
+
     @Override
     protected AbstractBeanDefinition parseInternal(final Element element, final ParserContext parserContext) {
-        BeanDefinitionBuilder factory = BeanDefinitionBuilder.rootBeanDefinition(ShardingTransactionTypeScanner.class);
+        BeanDefinitionBuilder factory = BeanDefinitionBuilder.rootBeanDefinition(ShardingAlgorithmFactoryBean.class);
+        factory.addConstructorArgValue(element.getAttribute(SPIBeanDefinitionParserTag.SHARDING_ALGORITHM_TYPE_ATTRIBUTE));
+        parseProperties(element, factory);
         return factory.getBeanDefinition();
     }
-    
-    @Override
-    protected boolean shouldGenerateId() {
-        return true;
+
+    private void parseProperties(final Element element, final BeanDefinitionBuilder factory) {
+        String properties = element.getAttribute(SPIBeanDefinitionParserTag.SHARDING_ALGORITHM_PROPERTY_REF_ATTRIBUTE);
+        if (!Strings.isNullOrEmpty(properties)) {
+            factory.addConstructorArgReference(properties);
+        } else {
+            factory.addConstructorArgValue(new Properties());
+        }
     }
 }
